// web/src/api/generated/models/ChapterDetail.ts
export interface ChapterDetail {
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof ChapterDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof ChapterDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof ChapterDetail
     */
    createdBy: string;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof ChapterDetail
     */
    creator: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof ChapterDetail
     */
    description: string | null;
    /**
     * 
     * @type {Date}
     * @memberof ChapterDetail
     */
    endOn: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ChapterDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ChapterDetail
     */
    placeId: string | null;
    /**
     * 
     * @type {string}
     * @memberof ChapterDetail
     */
    spaceId: string;
    /**
     * 
     * @type {Date}
     * @memberof ChapterDetail
     */
    startOn: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ChapterDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof ChapterDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof ChapterDetail
     */
    version: number;
}